Move FFI part for datetimes into its own module
